Move prometheus protobuf definitions into prometheus-parser lib
