+++ Cargo.toml
@@ -531,7 +531,7 @@ sinks-logdna = ["bytesize"]
 sinks-loki = ["bytesize"]
 sinks-nats = ["nats"]
 sinks-new_relic_logs = ["bytesize", "sinks-http"]
-sinks-prometheus = ["snap", "sources-utils-tls"]
+sinks-prometheus = ["prometheus-parser", "snap", "sources-utils-tls"]
 sinks-sematext = ["sinks-elasticsearch", "sinks-influxdb"]
 sinks-socket = ["sinks-utils-udp"]
 sinks-papertrail = []
