+++ src/lib.rs
@@ -52,8 +52,6 @@ pub mod list;
 pub mod mapping;
 pub mod metrics;
 pub(crate) mod pipeline;
-#[cfg(any(feature = "sinks-prometheus", feature = "sources-prometheus"))]
-pub(crate) mod prometheus;
 #[cfg(feature = "rusoto_core")]
 pub mod rusoto;
 pub mod serde;
